migrate from the workflow workaround to using the snyk API directly
